affichage message d'erreur modif rubrique

+++ spi-dosi/src/main/java/com/dosi/repository/RubriqueRepository.java
@@ -12,7 +12,7 @@ import com.dosi.models.Rubrique;
 public interface RubriqueRepository extends CrudRepository<Rubrique, Long>{
 	
 	@Query(value = "SELECT rubrique.* from rubrique,rubrique_question,rubrique_evaluation WHERE rubrique.ID_RUBRIQUE= ?1 AND rubrique.ID_RUBRIQUE=rubrique_question.ID_RUBRIQUE OR rubrique.ID_RUBRIQUE= ?1 AND rubrique.ID_RUBRIQUE=rubrique_evaluation.ID_RUBRIQUE", nativeQuery = true)
-	List<Rubrique> findIfIdRubriqueExistsInRbEvalRbQst(int idRubrique);
+	List<Rubrique> findIfIdRubriqueExistsInRbEvalRbQst(long idRubrique);
 	
 	
 	@Query(value = "SELECT rubrique.* from rubrique where rubrique.designation = ?1",nativeQuery = true)
